Add support for reauthentication against OIDC to activate sudo_mode
Currently, if the sudo_mode is enabled and the user tries to perform any
sensitive action (e.g. change application settings), they have to enter
their *local* password. However, if the user is logged in via OIDC, they
typically don't know their local password (it has been autogenerated on
the first login via OIDC).

This commit implements an opt-in feature that replaces this local
password verification with reauthentication on the OIDC provider, i.e.
even if the user is still logged in, they are forced to re-enter their
password (or any other kind of authentication) on the OIDC provider.